refactor: move NewUserRole to domain and complete MockDB impls
