+++ crates/router/src/db/kafka_store.rs
@@ -2575,7 +2575,7 @@ impl RedisConnInterface for KafkaStore {
 impl UserRoleInterface for KafkaStore {
     async fn insert_user_role(
         &self,
-        user_role: user_storage::NewUserRole,
+        user_role: domain::NewUserRole,
     ) -> CustomResult<user_storage::UserRole, errors::StorageError> {
         self.diesel_store.insert_user_role(user_role).await
     }
